Asteroids

+++ src/general/menu.h
@@ -8,12 +8,12 @@ class Menu
 {
     private:
         bool _firsttime = true;
-        int _currentIndex;
-        int _lastCounter = 0;
+        int16_t _currentIndex;
+        int16_t _lastCounter = 0;
         TFT_eSPI _tft;
         RotaryEncoder *_player1paddle;
 
-        int _gamecount = 4;
+        int16_t _gamecount = 4;
         String _items[4] = { "Asteroids", "1 Player Pong","Breakout", "Space Invaders" };
 
     public:
